+++ GuessWhoContracts/Dtos/RequestAndResponse/StartMatchRequest.cs
@@ -5,6 +5,6 @@ namespace GuessWhoContracts.Dtos.RequestAndResponse
     [DataContract]
     public class StartMatchRequest
     {
-        [DataMember(IsRequired = true)] public int MatchId { get; set; }
+        [DataMember(IsRequired = true)] public long MatchId { get; set; }
     }
 }
